+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - app information is now available through an "Info" button
 - app extra actions are now available through a "Settings" button instead of right-click
 - a confirmation popup is shown when the "install" / "upgrade all" button is clicked
+- "About" button added to the manage panel
+- "Updates" icon left-aligned
 - minor GUI improvements
 ### Fixes:
 - apps table not showing empty descriptions